refactor: reorganize Chatbar and Sidebar components into shared directory for better structure
